TR-1750 Moved selectDefaultTemplateOptions selector to account and added rest_tracking_default as default for click_tracking
