+++ src/selectors/tests/accessConditionState.test.js
@@ -1,7 +1,4 @@
-import selectAccessCondtionState, {
-  selectCondition,
-  selectDefaultTemplateOptions
-} from '../accessConditionState';
+import selectAccessCondtionState, { selectCondition } from '../accessConditionState';
 
 jest.mock('src/helpers/conditions/account');
 
@@ -13,11 +10,6 @@ describe('Selector: Access Condition State', () => {
   beforeEach(() => {
     testState = {
       account: {
-        options: {
-          click_tracking: true,
-          rest_tracking_default: true,
-          transactional_default: false
-        },
         subscription: {
           code: 'plan2'
         }
@@ -50,13 +42,5 @@ describe('Selector: Access Condition State', () => {
     const testCondition = jest.fn(() => true);
     expect(selectCondition(testCondition)(testState)).toEqual(true);
     expect(testCondition).toHaveBeenCalledWith(testAccessConditionState);
-  });
-
-  test('returns default template options', () => {
-    expect(selectDefaultTemplateOptions(testState)).toEqual({
-      click_tracking: true,
-      open_tracking: true,
-      transactional: false
-    });
   });
 });
